Fix FBO on OSX
