+++ video/screenshot.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 
 	"github.com/disintegration/imaging"
-	"github.com/go-gl/gl/all-core/gl"
+	"github.com/go-gl/gl/v2.1/gl"
 
 	"github.com/libretro/ludo/settings"
 	"github.com/libretro/ludo/state"
