+++ video/video.go
@@ -5,12 +5,11 @@ package video
 
 import (
 	"log"
-	"runtime"
 	"strconv"
 	"strings"
 	"unsafe"
 
-	"github.com/go-gl/gl/all-core/gl"
+	"github.com/go-gl/gl/v2.1/gl"
 	"github.com/go-gl/glfw/v3.3/glfw"
 	"github.com/go-gl/mathgl/mgl32"
 	"github.com/kivutar/glfont"
@@ -106,70 +105,6 @@ func getGLSLVersion() uint {
 	return uint(v)
 }
 
-// InitFramebuffer initializes and configures the video frame buffer based on
-// informations from the HWRenderCallback of the libretro core.
-func (video *Video) InitFramebuffer() {
-	width := video.Geom.MaxWidth
-	height := video.Geom.MaxHeight
-
-	log.Printf("[Video]: Initializing HW render (%v x %v).\n", width, height)
-
-	gl.GenFramebuffers(1, &video.fboID)
-	gl.BindFramebuffer(gl.FRAMEBUFFER, video.fboID)
-
-	//gl.GenTextures(1, &video.texID)
-	gl.BindTexture(gl.TEXTURE_2D, video.texID)
-	gl.TexStorage2D(gl.TEXTURE_2D, 1, gl.RGBA8, int32(width), int32(height))
-
-	gl.FramebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, video.texID, 0)
-
-	// Default origin is top left
-	video.orthoMat = mgl32.Ortho2D(-1, 1, -1, 1)
-
-	hw := state.Global.Core.HWRenderCallback
-
-	if hw != nil {
-		if hw.Depth && hw.Stencil {
-			gl.GenRenderbuffers(1, &video.rboID)
-			gl.BindRenderbuffer(gl.RENDERBUFFER, video.rboID)
-			gl.RenderbufferStorage(gl.RENDERBUFFER, gl.DEPTH24_STENCIL8, int32(width), int32(height))
-
-			gl.FramebufferRenderbuffer(gl.FRAMEBUFFER, gl.DEPTH_STENCIL_ATTACHMENT, gl.RENDERBUFFER, video.rboID)
-		} else if hw.Depth {
-			gl.GenRenderbuffers(1, &video.rboID)
-			gl.BindRenderbuffer(gl.RENDERBUFFER, video.rboID)
-			gl.RenderbufferStorage(gl.RENDERBUFFER, gl.DEPTH_COMPONENT16, int32(width), int32(height))
-
-			gl.FramebufferRenderbuffer(gl.FRAMEBUFFER, gl.DEPTH_ATTACHMENT, gl.RENDERBUFFER, video.rboID)
-		}
-
-		if hw.Depth || hw.Stencil {
-			gl.BindRenderbuffer(gl.RENDERBUFFER, 0)
-		}
-
-		if hw.BottomLeftOrigin {
-			video.orthoMat = mgl32.Ortho2D(-1, 1, 1, -1)
-		}
-	}
-
-	gl.BindRenderbuffer(gl.RENDERBUFFER, 0)
-
-	if gl.CheckFramebufferStatus(gl.FRAMEBUFFER) != gl.FRAMEBUFFER_COMPLETE {
-		log.Fatalln("[Video] Framebuffer is not complete.")
-	}
-
-	gl.ClearColor(0, 0, 0, 1)
-	if hw.Depth && hw.Stencil {
-		gl.Clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT | gl.STENCIL_BUFFER_BIT)
-	} else if hw.Depth {
-		gl.Clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT)
-	} else {
-		gl.Clear(gl.COLOR_BUFFER_BIT)
-	}
-
-	gl.BindFramebuffer(gl.FRAMEBUFFER, 0)
-}
-
 // Configure instanciates the video package
 func (video *Video) Configure(fullscreen bool) {
 	var width, height int
@@ -183,20 +118,6 @@ func (video *Video) Configure(fullscreen bool) {
 	} else {
 		width = 320 * 3
 		height = 180 * 3
-	}
-
-	// On OSX we have to force a core profile to not end up with 2.1 which cause
-	// a font drawing issue
-	if runtime.GOOS == "darwin" {
-		glfw.WindowHint(glfw.ContextVersionMajor, 3)
-		glfw.WindowHint(glfw.ContextVersionMinor, 2)
-		glfw.WindowHint(glfw.OpenGLProfile, glfw.OpenGLCoreProfile)
-		glfw.WindowHint(glfw.OpenGLForwardCompatible, glfw.True)
-	} else {
-		glfw.WindowHint(glfw.ContextVersionMajor, 2)
-		glfw.WindowHint(glfw.ContextVersionMinor, 1)
-		glfw.WindowHint(glfw.OpenGLProfile, glfw.OpenGLAnyProfile)
-		glfw.WindowHint(glfw.OpenGLForwardCompatible, glfw.False)
 	}
 
 	var err error
@@ -453,7 +374,7 @@ func (video *Video) ResizeViewport() {
 // Render the current frame
 func (video *Video) Render() {
 	// Render directly to the screen
-	gl.BindFramebuffer(gl.FRAMEBUFFER, 0)
+	bindBackbuffer()
 
 	// We can't trust the core to leave the OpenGL in the same state as
 	// before retro_run() was called so we restore some state manually.
@@ -510,7 +431,7 @@ func (video *Video) Render() {
 
 // Refresh the texture framebuffer
 func (video *Video) Refresh(data unsafe.Pointer, width int32, height int32, pitch int32) {
-	gl.BindFramebuffer(gl.FRAMEBUFFER, 0)
+	bindBackbuffer()
 
 	video.width = width
 	video.height = height
